Updates to Stages 5, 2 and 1
I've made some updates to accomodate the new location of the docket-sheet files as well as how the code in stage 2 selects the key ngrams.  These changes are focused on how the average is being calculated, the threshold for the minimum appearance and the maximum number of words that can be selected.